[WORK IN PROGRESS] Added Retrofit Network Implementation.
MealAPI class defines the URL endpoints of the MealDB.

RetrofitNetworkManager implements the methods contained in the API but I don't understand where/how/if the deserialization is done.

The MealDB returns the JsonArray "meals" which is then mapped to MealsContainer. I am able to get the meals but the networking and UI operations are still strongly coupled with one another and I still haven't find a solution to elegantly decouple them.